Treat client-side compilation as AOT loads
The initial implementation of the client code assumed that the compilations
it performs are AOT compilation until the remote request is issued and AOT
loads after the remote reply is received. This creates some confusion in
the code.
After this commit, a client-side compilation will be treated as an AOT load
from the beginning. The code determines if the server will be able to
perform a remote AOT compilation and only in that case a remote request is
issued. This filters out requests for JNI methods, thunks and methods that
are otherwise not AOT eligigle.

Another change is that the client is allowed to perform locally any compilation
request that cannot be done remotely. Moreover, if an AOT eligible compilation
request fails during relocation (maybe because classes it relies on at the
client side are not already loaded) it can be retried as a local JIT
compilation.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>