Change the protocol to a bidi-stream
To facilitate two-way communication within the scope of a compilation
request, the communication channel between the server and the client is
now a bidirectional stream. The server stream is asynchronous but all
reads and write are handled in a synchronous manner AFTER the stream
has been processed by the notification queue. The client stream is
synchronous as before. The server streams do not accept new RPCs until
their current one is completely processed; the notification queue can
hold outstanding connections in the interim.

Signed-off-by: Omer Sheikh <Omer.Sheikh@ibm.com>