Initial handling of server information requests
Process messages at the client on its stream in a loop until the server
sends back a compliationErrorCode. This should allow handling a variety
of information requests in the future.

Rename CompilationClient to J9ClientStream and J9CompileStream to
J9ServerStream to make it clear that these are counterparts of each
other.

Signed-off-by: Omer Sheikh <Omer.Sheikh@ibm.com>