Streamline the process of queueing a compilation in the server
To queue a compilation the JAAS server will call the newly created
TR::CompilationInfo::compileRemoteMethod()  which is a simplified version of
compileMethod() and compileOnSeparateThread() combined. More importantly,
it eliminates queries which could be answered only by the client.

The code assumes that compilations will be performed asynchronously and
that no requests for thunks, DLT and other non-ordinary methods will occur.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>