Change compilationThread infra to deal with remote requests
This commit enables the compilationThread infrastructure to deal with
compilation requests that come from a remote process.
It makes use of the newly introduced RemoteMethodDetails class and
in some instances it calls APIs that look at romClass/romMethod
(which are stored in RemoteMethodDetails) instead of j9method which
is a remote pointer that should never be dereferenced.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>