Allocate the exception table contiguously
The exeption table is expected to be allocated contiguously to have a
correct exception table offset when performing an AOT compilation.

This may be a temporary solution. Rewriting so that the parent class of
TR_ResolvedJ9JAASServerMethod is TR_ResolvedRelocatableJ9Method may be
a better solution, as that class is written to expect a contiguous
exception table.

Signed-off-by: Omer Sheikh <Omer.Sheikh@ibm.com>