Do not send unused values in mirrorResolvedJ9Method reply
The reponse to mirrorResolvedJ9Method would, in addition to creating a
client copy of the J9Method, also send the ROMClass, method name and
method signature to the server for the server to create its own copy.
The server does not yet copy over these values as it already has them.

Signed-off-by: Omer Sheikh <Omer.Sheikh@ibm.com>