Tag remote pointers with 0x7AA5
https://en.wikipedia.org/wiki/X86-64#Virtual_address_space_details
Since the x86-64 virtual address space is only 48 bits wide,
we can tag arbitrary 16 bit data on the highest bits of a remote pointer.
This creates a non-canonical pointer, which will cause the
processor to unconditionally raise an exception if dereferenced.
This is preferred over flipping all bits, because that makes pointer
arithmetic break.

Signed-off-by: Noah Weninger <noah.weninger@ibm.com>