Fix reported compilation time on the JAAS client
The JAAS client performs AOT load at this point (milestone 2).
However, the time needed to perform a remote compilation at the client
is reported wrongly in the verbose log, because we are missing a line
that sets the time when the compilation started.
For normal AOT loads the starting time is set at the very begining of the
compilation. For JAAS we determine much later that we will be doing
an AOT load, so we must set the compilation start time in that case too.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>