Disable late attach to self by default
Prevent a JVM from late attaching to itself unless a system property is set
either to "true" or an empty string, per the reference implementation.  Enforce
this in Java 9 only. Update existing tests to enable this feature by default so
they will continue to run.  Add new Java 9 tests to exercise this feature.

Signed-off-by: Peter Bain <peter_bain@ca.ibm.com>