Create initial .travis.yml to enable travis.ci builds
The initial travis yml is based off the Eclipse OMR .travis.yml and the
the OpenJ9 docker file.  This does the necessary apt-gets and wgets to
create the build environment.

The build runs get_sources.sh to set up the 3 required repos.  As
Travis has already cloned the openj9 repo locally, we pass that
copy of the repo's location, sha, and branch to the get_sources.sh
script.

This results in a duplicate copy of the openj9 repo on disk but avoids
network traffic - overall a win.

The build ran into some strange issues:

* the bootjdk detection fails unless the '_JAVA_OPTIONS=' env var is
unset

* the compiles fail with a compiler internal error unless the ojdk
JOBS are limited and the `UMA_WINDOWS_PARRALLEL_HACK` is used to
control the `-jN`.  We can't directly set `-jN` in the MAKE env as
the OJDK build scripts fail when they detect it.  Using -j4 as the
travis container build env is 2 core / 4 gig memory.

* ccache is also disabled as the ojdk build scripts don't support
using a wrapped compiler.  They generate the following error:
   configure: error: /usr/lib/ccache/gcc is a symbolic link to ccache.
              This is not supported.
   configure exiting with result code 1

TODOs:
* Re-enable ccache
* Add testing to this build

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>