Send the ROM class and method over the wire
Instead of reading the rom class and method from the shared class cache
when creating a TR_ResolvedJ9JAASServerMethod, we now receive the bytes
for the rom structures from the client directly. Self referential
pointers to the method name and signature are manually updated on the
client side to allow the ROMClass to be sent over as a single block
of data. There are still SRPs to data that may lie outside of the class
and need to be fixed.

There are still places where the code reads the rom class and method
from the shared class cache; including at the start of a compilation.

Signed-off-by: Omer Sheikh <Omer.Sheikh@ibm.com>