Add a thread_local pointer to CompInfoPerThread
Adds a thread_local `TR_CompilationInfoPerThread *` variable to the TR
namespace for storing an unordered map of the client J9Classes to server
J9ROMClasses and for providing access to the JAAS stream object. Also
removes the thread_local stream pointer in `TR::CompilationInfo` as it
is no longer necessary with this change.

Signed-off-by: Omer Sheikh <Omer.Sheikh@ibm.com>