Server generates compiled body at address indicated by client
Notes:
1) Code cache repositories at the server and client need to have the
same address. This has a bigger chance of happening if Linux ASLR is disabled
with  `sudo sysctl kernel.randomize_va_space=0`
Moreover, the client can request a specific address where code cache
repository is loaded by using -Xjit:mandatoryCodeCacheAddress=...

2) To run bigger applications the size of a code cache needs to be increased
with -Xjit:code=...  For Liberty count=1 I had success with -Xjit:code=16384
Otherwise we exhaust the cache and the server throws an exception which in
theory should propagate to the client to allocate a new code cache. This
mechanism is missing today.

3) It's safer to allocate all the code caches at the server by using
numCodeCachesOnStartup=16 (assuming the entire repository is 256M and the
size of one code cache is 16M as suggested above)

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>