Fix casting issues in generateRILInstruction() API
Splitting the generateRILInstruction() API to handle the different
types of immediates that RIL instructions can take separately.
One kind is a pure immediate which is used for say an ADD
instruction. The other is when the immediate is used as a relative
offset for address calculation like LARL or BRASL instructions.
This allows calls to the API to have more sensible casting.
They are modified for this change and unsafe casting is switched
to static casting and some code has its style improved.

Signed-off-by: Simon Hirst <shirst@ca.ibm.com>