Correct escape analysis loop check
Escape analysis may have to create initialization
code at the method start. If the first block is
part of a loop, it will create a new block to hold
these trees. However, the block test did not consider
whether the first block was in an improper region,
resulting in init code being placed inside a loop.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>