Simplify check for class initialization
Side effect guards are runtime guards that can
be patched on class initialization. When commiting
the guard, it is necessary to check such initialization
has not yet occured. The prior check would walk the
entire table of classes to determine if one of the
assumed was present. However, there are few classes
where this assumption is made, so it is cheaper to
query them directly.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>