Cache CFG end in OSRGuardInsertion
OSRGuardInsertion will introduce several OSR guards
during a compilation, attaching the taken side
to the end of the CFG. In the past, this required
repeated walks of the CFG to find the end. The
OSR induce API has been updated to take a treetop
where the taken side should be inserted after and
will update it to point at the end of the added
block.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>