Skip classes with null jniIDs field when updating methodIDs
During a hotswap, fixJNIMethodID() scans the list of replaced classes searching
for the old methodID and updates it, until it finds a class without the old
methodID.  In some cases, a class's list of jniIDs may be null.  To prevent a
crash, do a null check before scanning the list.

Signed-off-by: Peter Bain <peter_bain@ca.ibm.com>