Add java.lang.string.substring.share System property
In some scenarios it may be undesirable to share the underlying value
array for substring operations as this transparently increases the life
of the respective array. For carefully crafted user applications this
can cause us to run out of heap memory when performing repeated
substring operations, e.x. repeatedly calling substring to extract parts
of a large string into an array.

This change introduces a system property which can be used to disable
sharing of the value array in substring operations, regardless of
whether the value array can be shared between the receiver string and
the resulting substring.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>