Add New Logo Colour Combinations
Here are three new colour combinations for the OpenJ9 logo.

OpenJ9-all-black.svg is an all black variant for use on light
backgrounds.
OpenJ9-all-white.svg is an all white variant for use on dark
backgrounds.
OpenJ9-white-J9.svg is a teal Open and circle with a white J9 that
works well on dark backgrounds.

[skip ci]

Signed-off-by: Ronald Servant <ronald_servant@ca.ibm.com>