Fix variable precision off by one issue and improve checking
ZAP instruction immediate length has an implied length of one if zero is
encoded in the instruction. As such we need to determine the number of
bytes from the precision minus one. 

In addition add TP instructions for pd2i and pd2l (as well as variable
precision) to check the contents of the packed decimal before executing
ZAP / CVB / CVBG. This is to ensure that if a zero initialized array is
passed through we directly branch to the OOL path to execute the Java
implementation rather than taking the signal handler path.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>