Add test packed decimal before PD2I conversions
Add TP or VTP instructions for pd2i and pd2l (as well as variable
precision) to check the contents of the packed decimal before executing
ZAP / CVB / CVBG. This is to ensure that if a zero initialized array is
passed through we directly branch to the OOL path to execute the Java
implementation rather than taking the signal handler path.

Also, fix a length-code off-by-1 bug. The length-code used by ZAP
instruction should be packed decimal byte-length minus 1. The AGHI
(for plus 1) should not be there.

Signed-off-by: Nigel Yu <yunigel@ca.ibm.com>