Make the main thread from JAAS server sleep after VM init
The goal of this change set is to be able to run the server with
no Java application required and no special options on the
command line. Thus,
   java -XX:jaasServer
should be enough.
The change set registers a hook that fires when the VM initialization
is complete (towards the end of J9_CreateJavaVM())
In this hook we place a sleep operation for a very long time.
The server must be stopped through a kill operation.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>