More optimized class initialization check in MethodHandle
Check if a class has been initialized by calling a JITHelpers method
which can be optimized by the JIT.
Original implemenation uses Unsafe to load the class flag from J9Class
and JIT cannot see through it. With this change, JIT is able to
eliminate the check if it knows what the class is and that the class
has been initialized.
This change also caches the JITHelpers object and the Unsafe object in
MethodHandle as a static field so that when a JITHelpers call is
reduced, the load of JITHelpers object can also be reduced.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>