arraytranslate for StringCoding.implEncodeISOArray
The implEncodeISOArray in java/lang/StringCoding is equivalent to the
old sun.nio.cs.ISO_8859_1.Encoding.encodeISOArray. This change adds
support for this new method and compensates for the different length of
the source array in the necessary safety checks. This change also
improves the symref used for StringUTF16.getChar for performance.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>