Improve reliability of travis builds
The frequent failures we observe suggest that the TRAVIS_COMMIT
variable cannot be relied upon to give the SHA of the merge commit
to be built. One theory is the it is given a value at the time the
job is queued which will be stale if any commits are pushed to either
the source or target branches.

The build clones the repo and checks out the merge commit by name.
This change recovers the full SHA from the local clone.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>