Re-enable StringCoding optimization
An issue was found in testing with recent changes to accelerate
StringCoding in Java9. The cause of the issue was a change to skip array
zero initialization. When running with arraylets enabled (eg GC modes
like balanced) the zero initialization is required). This guards the
zero initialization skipping with a test for arraylets.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>