Fix segfault in insertGPUTemporariesLivenessCode
The variable currentNode inside insertGPUTemporariesLivenessCode can
sometimes be used without being initialized. This can lead to problems such
as a segfault.

This change updates insertGPUTemporariesLivenessCode to ensure that
currentNode is initialized to a useable value before being used. It also
cleans up the code in the function to remove unused variables and to make
the code logic easier to follow.

Issue: #557
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>