Compressed Reference Support for Loads
Auto SIMD will no longer invalidate an entire loop for vectorization if a
compressed reference load is present. Auto SIMD will ignore this treetop, and
continue to determine if any other tree tops in the loop are candidates for
vectorization. If the load is present elsewhere in the loop, it will be
vectorized later.

This approach cannot be taken with a compressed reference tree top with a
store child. Ignoring the compressed reference treetop will lead to an
unvectorized store within a loop that will be vectorized.

Signed-off-by: Ryan Santhirarajan <rsanth@ca.ibm.com>