Prefetch, checkcast and profiled compile changes
HashMap.resize and HashMapHashIterator.nextNode were added to
canSkipCheckCasts since they don't need check casts.

Node threshold inside ProfileGenerator was changed to 90000. This will
promote more profiled compilations and was shown to improve the throughput
of an internal appserver benchmark.

addPrefetch was moved from OMR to OpenJ9 and renamed to
insertPrefetchIfNecessary.

It was also modified to work on Itr.next in ArrayList and the prefetching
stride distance was also changed. The new value was found through testing
and displayed an improvement in throughput in the previously mentioned
internal appserver benchmark.

Also during the move, an bug in insertPrefetchIfNecessary where
compressedrefs decompression was not occuring was fixed.

iwrtbarEvaluator was modified to now use insertPrefetchIfNecessary instead
of addPrefetch.

Issue: #2079

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>