Enable jck test signaturetest to pass
JCK test "test.jck8b.runtime.api.signaturetest" fails because the list
of method signatures it has does not reflect reality in our OpenJDK8 on
OpenJ9 build. One of the J9JCL MethodHandle methods does not exist in
OpenJDK8/9/10 MethodHandle, and another has the wrong scope. I've
altered a signature and removed a method to match.

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>