Re-enable sub-tests in URLStaleClassPathEntryTest
1. Re-enable URLStaleClassPathEntryTest 1-4 on Java 9
2. Jar file has been updated after classes are stored to the shared
cache, we should not be able to find the classes from the shared cache.
The expected test results of URLStaleClassPathEntryTest 4 in Test4.props
should be modified. 
3. Add "cache does not exist" to the success condition in
URLHelperTests.xml when cleaning up the shared cache from the previous
tests. We do not need to care about the destroy message as long as the
shared cache is deleted.

Fixes #470

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>