Fixup forwarded referent slot
In case of an aborted Concurrent Scavenge, in follow-up Percolate Global
GC, when doing final clearing pass, if encountering a referent slot
(within any of soft/weak/phantom refs), that already points to a
copy&forwarded object, fixup up the referent slot to point to the
forwarded version of the object. Then liveness of the referent would be
verified on the forwarded version.

We used to do it only for phantom reference, which would lead to
incorrectly clearing weak/soft reference (since liveness would be
checked on the  original version of the referent which would not be
marked, even though forwarded version would).

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>