Relocate reserveNTrampolines to x86 CodeGenerator
Move the `reserveNTrampolines()` implementation from the FrontEnd to
the x86 CodeGenerator where it more naturally belongs.  The API is
also simplified to only accept the number of trampolines to reserve.
The rest of the original API can be derived from other queries
on the `TR::CodeGenerator` object.

Nothing uses this function yet.  A subsequent commit will delete the
implementation in the FrontEnd and re-route all references to the
version in CodeGenerator.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>