Remove unnecessary files for building JitBuilder
While building JitBuilder on OS/X, the archiver warns about
files which it believes to be responsible for generating empty
objects.

Upon closer inspection, that indeed seems to be the case.

The actual files are being removed in eclipse/omr/pull/1740.
This commit acts as a paired dependency change on the OpenJ9 side.

Signed-off-by: Rwitaban (Ray) Banerjee <rayb@ca.ibm.com>