Migrate J9 nodeNeeds2Regs from Compilation to Node
This is the OpenJ9-specific change for eclipse/omr#1886, which has
already been merged in Eclipse OMR.

A node should be responsible for answering whether it requires a register
pair or not.

Move the functionality to the Node class, rename the function to
`requiresRegisterPair`, and add some documentation.

Also, remove the `nodeNeeds2Regs` function from the J9 Compilation class.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>