Use libj9prtstatic.a instead of libj9omrport.a
OpenJ9's port library libj9prt29.so depends on OMR's port library.
During creation of libj9prt29.so, OpenJ9 also compiles OMR
port library code to create a static library libj9omrport.a.
However, OMR already creates a static library libj9prtstatic.a which
is same as libj9omrport.a created by OpenJ9.
It would be better if libj9prt29.so is linked against libj9prtstatic.a.
This allows libj9prt29.so to be updated by 'make' whenever any change
is made in OMR port library code and removes the need to explicitly
create libj9omrport.a in such case.

Signed-off-by: Ashutosh Mehra <asmehra1@in.ibm.com>