Stop allocating memory from already released segments
A large system segment, one that's larger than a specific size, is
used by only one region and is to be release once that region is out
of scope.

A field in J9::SystemSegmentProvider named _currentSystemSegment keeps
track of current active segment from where allocation can be made.
This field will point to a large system segment if one is created.
When this segment is released, this field is not updated, resulting
potential allocations from such segment, then causing segmentation
errors due to read/write to inaccessible addresses. The inaccessible
address usually appears to a seemingly valid one.

This change stops _currentSystemSegment holding onto such segment.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>