Remove exception throw from processor detection
Processor detection occurs at initialization time, and therefore,
outside of a compilation. Additionally, it doesn't make sense to guard
this code in a try/catch since unsupported processors are a fatal error.
Therefore, this commit removes the throwing of exceptions and replaces
the TR_ASSERT with a TR_ASSERT_FATAL.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>