Fix computed dispatchJ9Method calls for JaaS
If Inlining of natives is enabled then compiler modifies the symbol
reference of TR::java_lang_invoke_ComputedCalls_dispatchJ9Method calls
to hold the address of the helper. When creating a relocation for such
calls, OMR does not know that the address is that of a helper and
creates a TR_RelativeMethodAddress relocation instead of a
TR_HelperAddress relocation. This change transmutes the symbol
reference to appear as a helper reference so that a TR_HelperAddress
relocation is generated instead.

Signed-off-by: Omer Sheikh <Omer.Sheikh@ibm.com>