Remove pdshlOverflow above the pdshr
The pdshlOverflow above pdhsr node is redundant and can be removed for the
following two reasons.

First, both the pdshr node and its parent pdshlOVerflow have the same decimal
precision. The psdhlOverflow has a shift amount of zero. So this overflow
node is unneeded. Second, the pdshr node is itself an
exceptioning node. Placing the pdshr under the BCDCHK should be legal.

Signed-off-by: Nigel Yu <yunigel@ca.ibm.com>