Call initializeTargetEnvironment() in specialized class
Call initializeTargetEnvironment() via OMR::CompilerEnvConnector
so that platform-specific initialization is performed.
In this case, it is making sure the function in
OMR::ARM::OMRCompilerEnv is called, so that setEABI() is called.
The code in the old initializeTargetEnvironment(), i.e., bitness,
CPU, and OS settings, in J9CompilerEnv.cpp has been already moved
to OMR::CompilerEnv class.

Closes: #619

Signed-off-by: knn-k <konno@jp.ibm.com>