Free the initial J9ROMclass for remote compiles
If the compilation entry is added and process, clean up the rom class
after clearing the rom classes cache so that it is freed regardless of
the path taken during the compilation.

If a remote compilation never reaches a compilation thread, still free
the rom class in the failure branches in doAOTCompile.

Signed-off-by: Omer Sheikh <Omer.Sheikh@ibm.com>