NegativeArraySizeException not thrown from StringBuffer constructor
NegativeArraySizeException is not thrown from the StringBuffer
constructor that takes capacity as an argument. In the constructor,
capacity argument is used to determine the byte/char array size. If
the capacity argument is Integer.MIN_VALUE (-2147483648), then capacity
*= 2 will yield a non-negative number due to overflow. Thus, we will
fail to throw NegativeArraySizeException. A new check has been added to
assure that NegativeArraySizeException is thrown if capacity argument is
less than 0.

The implementation of similar constructor in StringBuilder has been
updated to maintain consistency between StringBuffer and StringBuilder.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>