Modify ValueProfiling APIs
This change attempts to cleanup the ValueProfiling
API. It achieves this by allowing for a type and
source to be specified when extracting values from
the API or inserting instrumentation.

It also removes functionality to access profiling
results from the mechanism for inserting profiling
instrumentation, ValueProfiler. Instead, profiling
information should either be accessed through
the TR_ValueProfileInfoManager or TR_ValueProfileInfo
directly.

Moreover, the concepts of the profiled value and
its profiling instrumentation have been separated,
so that a recompilation can inspect information
regardless of how it was collected.

Finally, heap allocations used for IProfiler
have been removed from TR_ValueProfileInfo, so that
its simple to distinguish between persistent
and heap structures.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>