Add Value JProfiling Implementation
This adds a new value profiling implementation,
capable of collecting information with low
overhead. It achieves this by only tracking
the most frequent values and stores them in
a hash map to simplify lookups. A majority
of the instrumentation is represented in
trees.

This is used in an optimization pass, capable
of identifying common profiling candidates in
arbitrary compilations, as well as in a modified
ValueProfiler, used to instrument profiling
compilations.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>